+++ app/src/main/java/com/starters/hsge/data/interfaces/UserInfoPutInterface.kt
@@ -1,0 +1,8 @@
+package com.starters.hsge.data.interfaces
+
+interface UserInfoPutInterface {
+
+    fun onPutUserInfoSuccess(isSuccess: Boolean, code: Int)
+
+    fun onPutUserInfoFailure(message: String)
+}
